Move tests to correct files (finish work with all rule exclude lenght)
