+++ detekt-rules-naming/src/test/kotlin/io/gitlab/arturbosch/detekt/rules/naming/FunctionNamingSpec.kt
@@ -4,7 +4,10 @@ import io.gitlab.arturbosch.detekt.api.SourceLocation
 import io.gitlab.arturbosch.detekt.test.TestConfig
 import io.gitlab.arturbosch.detekt.test.assertThat
 import io.gitlab.arturbosch.detekt.test.compileAndLint
+import org.assertj.core.api.Assertions
+import org.junit.jupiter.api.Nested
 import org.junit.jupiter.api.Test
+import java.util.regex.PatternSyntaxException
 
 class FunctionNamingSpec {
 
@@ -108,5 +111,75 @@ class FunctionNamingSpec {
             fun `7his is a function name _`() = Unit
         """
         assertThat(FunctionNaming().compileAndLint(code)).hasStartSourceLocations(SourceLocation(1, 5))
+    }
+
+    @Test
+    fun `should use custom name for method`() {
+        val config = TestConfig(mapOf(FunctionNaming.FUNCTION_PATTERN to "^`.+`$"))
+        assertThat(
+            FunctionNaming(config).compileAndLint(
+                """
+        class Foo {
+            fun `name with back ticks`(){
+              
+            }
+        }
+                """
+            )
+        ).isEmpty()
+    }
+
+    @Test
+    fun shouldExcludeClassesFromFunctionNaming() {
+        val code = """
+        class Bar {
+            fun MYFun() {}
+        }
+
+        object Foo {
+            fun MYFun() {}
+        }
+        """
+        val config = TestConfig(mapOf(FunctionNaming.EXCLUDE_CLASS_PATTERN to "Foo|Bar"))
+        Assertions.assertThat(FunctionNaming(config).compileAndLint(code)).isEmpty()
+    }
+
+    @Test
+    fun `should report a function name that begins with a backtick, capitals, and spaces`() {
+        val subject = FunctionNaming()
+        val code = "fun `Hi bye`() = 3"
+        subject.compileAndLint(code)
+        Assertions.assertThat(subject.findings).hasSize(1)
+    }
+
+    @Nested
+    inner class `exclude class pattern function regex code cases` {
+        private val excludeClassPatternFunctionRegexCode = """
+            class Bar {
+                fun MYFun() {}
+            }
+
+            object Foo {
+                fun MYFun() {}
+            }
+    """
+
+        @Test
+        fun shouldNotFailWithInvalidRegexWhenDisabledFunctionNaming() {
+            val configRules = mapOf(
+                "active" to "false",
+                FunctionNaming.EXCLUDE_CLASS_PATTERN to "*Foo"
+            )
+            val config = TestConfig(configRules)
+            Assertions.assertThat(FunctionNaming(config).compileAndLint(excludeClassPatternFunctionRegexCode)).isEmpty()
+        }
+
+        @Test
+        fun shouldFailWithInvalidRegexFunctionNaming() {
+            val config = TestConfig(mapOf(FunctionNaming.EXCLUDE_CLASS_PATTERN to "*Foo"))
+            Assertions.assertThatExceptionOfType(PatternSyntaxException::class.java).isThrownBy {
+                FunctionNaming(config).compileAndLint(excludeClassPatternFunctionRegexCode)
+            }
+        }
     }
 }
