+++ detekt-rules-naming/src/test/kotlin/io/gitlab/arturbosch/detekt/rules/naming/PackageNamingSpec.kt
@@ -1,10 +1,17 @@
 package io.gitlab.arturbosch.detekt.rules.naming
 
+import io.gitlab.arturbosch.detekt.test.TestConfig
 import io.gitlab.arturbosch.detekt.test.compileAndLint
 import org.assertj.core.api.Assertions.assertThat
 import org.junit.jupiter.api.Test
 
 class PackageNamingSpec {
+
+    @Test
+    fun `should use custom name for package`() {
+        val rule = NamingRules(TestConfig(PackageNaming.PACKAGE_PATTERN to "^(package_1)$"))
+        assertThat(rule.compileAndLint("package package_1")).isEmpty()
+    }
 
     @Test
     fun `should ignore the issue by alias suppression`() {
